+++ packages/gamut-labs/src/GlobalFooter/FooterNavLinks/CompanyLinks.tsx
@@ -4,7 +4,6 @@ import React from 'react';
 
 import { LogoFromSkillsoft } from '../..';
 import { footerResourcesList } from '../../lib/resourcesList';
-import { FooterFFContext } from '..';
 import { FooterHeading } from '../FooterHeading';
 import {
   FooterLinkItem,
@@ -95,17 +94,7 @@ export const CompanyLinks: React.FC<CompanyLinksProps> = ({
   const company = (
     <Box>
       <FooterHeading>
-        <FooterFFContext.Consumer>
-          {({ showNewFooter }) =>
-            showNewFooter ? (
-              <Box>
-                <LogoFromSkillsoft height={40} />
-              </Box>
-            ) : (
-              'Company'
-            )
-          }
-        </FooterFFContext.Consumer>
+        <LogoFromSkillsoft height={40} />
       </FooterHeading>
       <FooterLinkItems>
         <FooterLinkItem>